fix printvalue template

+++ lifter/utils.cpp
@@ -74,7 +74,7 @@ namespace debugging {
     template void printValue<unsigned long>(const unsigned long& v,
                                             const char* name);
     template void printValue<long>(const long& v, const char* name);
-    template void printValue<__int64>(const __int64& v, const char* name);
+    template void printValue<long long>(const long long& v, const char* name);
     template void printValue<KnownBits>(const KnownBits& v, const char* name);
     template void printValue<ROP_info>(const ROP_info& v, const char* name);
     template void printValue<unsigned long long>(const unsigned long long& v,
